Move current selector to belong only to compensation
